requirements.txt
Django==3.2.14
djangorestframework==3.13.1
djoser==2.1.0
python-dotenv==1.0.1
psycopg2-binary==2.9.5